renamed server to service
